+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* New Tag filter `∃key∉value1,value2,value3,…`
+
 ## v3.6.0 (2026-01-08)
 
 * Minor bugfixes